Refactoring mjai.bot
