+++ python/mjai/bot/__init__.py
@@ -1,0 +1,1 @@
+from mjai.bot.base import Bot  # noqa: F401
